refactor(build-logic): switch to precompiled buildscripts
