+++ build.gradle.kts
@@ -15,6 +15,5 @@
  */
 plugins {
     base
-    alias(baseLibs.plugins.kotlin.jvm) apply false
-    alias(baseLibs.plugins.maven.gradle.publish.plugin) apply false
+    alias(baseLibs.plugins.maven.gradle.publish) apply false
 }
