+++ version-catalog/libs.versions.toml
@@ -1,7 +1,6 @@
 [versions]
 # Kotlin and Gradle
 kotlin = "2.2.20"
-kotlin-dsl = "6.4.0"
 jetbrains-annotations = "26.0.2-1"
 plugin-publish = "2.0.0"
 
@@ -51,7 +50,7 @@ assertj = "4.0.0-M1"
 spock-core = "1.3-groovy-2.4"
 
 # Publishing
-maven-gradle-publish-plugin = "0.34.0"
+maven-gradle-publish = "0.34.0"
 
 # Miscellaneous
 jgit = "7.4.0.202509020913-r"
@@ -114,10 +113,10 @@ jgit-gradle-plugin = { module = "org.eclipse.jgit:org.eclipse.jgit", version.ref
 
 [plugins]
 kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
-kotlin-dsl = { id = "org.gradle.kotlin.kotlin-dsl", version.ref = "kotlin-dsl" }
 kover = { id = "org.jetbrains.kotlinx.kover", version.ref = "kover" }
 owaspDependencyCheck = { id = "org.owasp.dependencycheck", version.ref = "owaspCheck" }
-versions = { id = "com.github.ben-manes.versions", version.ref = "versions" }
 dokka = { id = "org.jetbrains.dokka", version.ref = "dokka" }
+spotless = { id = "com.diffplug.spotless", version.ref = "spotless" }
+detekt = { id = "io.gitlab.arturbosch.detekt", version.ref = "detekt" }
 plugin-publish = { id = "com.gradle.plugin-publish", version.ref = "plugin-publish" }
-maven-gradle-publish-plugin = { id = "com.vanniktech.maven.publish", version.ref = "maven-gradle-publish-plugin" }
+maven-gradle-publish = { id = "com.vanniktech.maven.publish", version.ref = "maven-gradle-publish" }
